+++ clap_up/src/lib.rs
@@ -5,9 +5,35 @@ use cargo_up::{
     upgrader, Upgrader, Visitor,
 };
 
-#[upgrader]
+#[upgrader(minimum = "2.33.0", peers = ["structopt"])]
 #[derive(Default)]
 pub struct Clap;
+
+// #[up("3.0.0-beta.1")]
+// replace_dep!("structopt", "clap", features = ["derive"]);
+
+// #[up("3.0.0-beta.1")]
+// rename_struct_methods!(clap::args::arg, Arg, [
+//     help => about,
+//     from_usage => from,
+//     set => setting,
+//     unset => unset_setting,
+// ]);
+
+// #[up("3.0.0-beta.1")]
+// rename_struct_methods!(clap::app, App, [
+//     from_yaml => from,
+//     arg_from_usage => arg,
+//     help => override_help,
+//     usage => override_usage,
+//     template => help_template,
+//     get_matches_safe => try_get_matches,
+//     get_matches_from_safe => try_get_matches_from,
+//     get_matches_from_safe_borrow => try_get_matches_from_mut,
+// ]);
+
+// #[up("3.0.0-rc.0")]
+// rename_struct_methods!(clap::build::arg, Arg, with_name => new);
 
 impl Visitor for Clap {
     fn visit_method_call_expr(
@@ -15,17 +41,22 @@ impl Visitor for Clap {
         method_call_expr: ast::MethodCallExpr,
         semantics: &Semantics<RootDatabase>,
     ) {
-        let f = semantics.resolve_method_call(&method_call_expr).unwrap();
-        let fn_name = format!("{}", f.name(semantics.db));
+        if let Some(name_ref) = method_call_expr.name_ref() {
+            if name_ref.text() != "help" {
+                return;
+            }
 
-        if let Some(name) = get_struct_name(&f, semantics.db) {
-            let mod_name = full_name(&f.module(semantics.db), semantics.db);
+            let f = semantics.resolve_method_call(&method_call_expr).unwrap();
 
-            if format!("{}", name) == "Arg" && mod_name == "clap::args::arg" && fn_name == "help" {
-                self.replace(
-                    method_call_expr.name_ref().unwrap().syntax().text_range(),
-                    "about".to_string(),
-                );
+            if let Some(name) = get_struct_name(&f, semantics.db) {
+                let mod_name = full_name(&f.module(semantics.db), semantics.db);
+
+                if format!("{}", name) == "Arg" && mod_name == "clap::args::arg" {
+                    self.replace(
+                        method_call_expr.name_ref().unwrap().syntax().text_range(),
+                        "about".to_string(),
+                    );
+                }
             }
         }
     }